Read numeric values with cells' DataFormatter
Although all numeric values in excel are stored as doubles, the
formatter string defines how they are presented (e.g. as integers
or with decimal places).

This patch allows us to distinguish numeric types when reading (with the
assumption that the choice of type is reflected in the choice of
presentation format).